move: gallery 도메인 엔티티들을 entity 패지키로 이동
