+++ src/main/java/org/younginhambak/backend/file/entity/PhotoFile.java
@@ -5,6 +5,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.younginhambak.backend.gallery.Photo;
+import org.springframework.util.Assert;
 
 import java.time.LocalDateTime;
 import java.util.EnumSet;
