--- src/main/java/org/younginhambak/backend/gallery/PhotoStatus.java
@@ -1,5 +1,0 @@
-package org.younginhambak.backend.gallery;
-
-public enum PhotoStatus {
-  ACTIVE, DISABLED, DELETED, ARCHIVED
-}
