+++ src/main/java/org/younginhambak/backend/gallery/entity/PhotoStatus.java
@@ -1,0 +1,5 @@
+package org.younginhambak.backend.gallery.entity;
+
+public enum PhotoStatus {
+  ACTIVE, DISABLED, DELETED, ARCHIVED
+}
