+++ src/main/java/org/younginhambak/backend/gallery/entity/PhotoTagId.java
@@ -1,0 +1,15 @@
+package org.younginhambak.backend.gallery.entity;
+
+import jakarta.persistence.Embeddable;
+import lombok.*;
+
+import java.io.Serializable;
+
+@Embeddable
+@Getter
+@EqualsAndHashCode
+@AllArgsConstructor
+public class PhotoTagId implements Serializable {
+  private Long photoId;
+  private Long tagId;
+}
